Update position of buttons
